put Config in external class
